pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">=3.12,<3.13"
name = "routetools"
description = "Routing Tools for Continuous Vector Fields"
readme = "README.md"
authors = [
    { name = "Weather Routing Research", email = "daniel.precioso@ie.edu" },
]
# For reproducible lockfiles and environments we keep a static version here.
version = "0.1.0"
dependencies = [
    "pandas",
    "typer-slim",
    "jax[cuda]==0.8.1",
    "jaxlib==0.8.1",
    "matplotlib>=3.9.2",
    "cma>=4.0.0",
    "pytest>=8.3.2",
    "shapely>=2.0.6",
    "xarray>=2024.11.0",
    "perlin-numpy>=0.0.1",
    "seaborn>=0.13.2",
    "typer>=0.19.2",
    "h3==4.0.0b2",
    "borb @ https://files.pythonhosted.org/packages/7a/4e/b193d894ffb0fde0f773b0476d8b041528302fa86bccf2cc8006c79404e4/borb-3.0.1.tar.gz",
    "jupyter-server>=2.17.0",
    "statsmodels>=0.14.6",
    "scipy>=1.12.0",
    "netCDF4>=1.6.0",
    "opencv-python>=4.13.0.92",
    "dask>=2024.1.0",
    "cartopy>=0.25.0",
    "gcsfs>=2026.2.0",
    "zarr>=3.1.5",
]

[project.optional-dependencies]
cuda = ["jax[cuda]==0.8.1"]
era5-cds = ["cdsapi>=0.7.0"]
era5-gcs = ["gcsfs>=2024.1.0", "zarr>=2.17.0"]
swopp3 = ["swopp3-performance-model==0.1.0"]

[project.urls]
homepage = "https://github.com/Weather-Routing-Research/cmaes_bezier_demo"

[project.scripts]
main = "routetools.main:main"

[tool.setuptools]
packages = [
    "routetools",
    "routetools.era5",
    "routetools._cost",
    "routetools.wrr_bench",
]

[tool.setuptools_scm]
# Configure setuptools_scm to write a version file and provide a fallback when
# git metadata is not available (this ensures editable installs and lockfile
# generation can see a concrete version).
write_to = "routetools/_version.py"
fallback_version = "0+unknown"

[tool.ruff]
line-length = 88
[tool.ruff.lint]
select = [
    "E",    # pycodestyle
    "F",    # Pyflakes
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",    # isort
    "D",    # pydocstyle
    "C401", # flake8-comprehensions: unnecessary-generator-set
    "C402", # flake8-comprehensions: unnecessary-generator-dict
    "C403", # flake8-comprehensions: unnecessary-list-comprehension-set
    "C404", # flake8-comprehensions: unnecessary-list-comprehension-dict
    "C405", # flake8-comprehensions: unnecessary-literal-set
    "W605", # pycodestyle: invalid-escape-sequence
]

# Ignore D100 Missing docstring in public module
ignore = ["D100"]
[tool.ruff.lint.per-file-ignores]
# Also ignore `D104` in all `__init__.py` files.
"__init__.py" = ["D104"]
"tests/*.py" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.mypy]
files = "./routetools"
strict = true
warn_unreachable = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]

[tool.pytest.ini_options]
minversion = "7"
testpaths = ["tests"]
log_cli_level = "INFO"
xfail_strict = true
addopts = ["-ra", "--strict-config", "--strict-markers"]
filterwarnings = ["error", "ignore::DeprecationWarning"]

[dependency-groups]
dev = ["pytest", "ruff", "mypy", "pre-commit", "ipykernel"]
